+++ applications/tile-inspector/CHANGELOG.md
@@ -1,5 +1,21 @@
 # tile-inspector
 
+## 0.1.63
+
+### Patch Changes
+
+- Updated dependencies [58d28355]
+  - @globalfishingwatch/layer-composer@8.10.0
+
+## 0.1.62
+
+### Patch Changes
+
+- Updated dependencies [0ea624c9]
+- Updated dependencies [08b8209f]
+  - @globalfishingwatch/layer-composer@8.9.1
+  - @globalfishingwatch/api-client@2.0.1
+
 ## 0.1.61
 
 ### Patch Changes