+++ applications/vessel-history/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @globalfishingwatchapp/vessel-history
 
+## 0.0.72
+
+### Patch Changes
+
+- Updated dependencies [58d28355]
+  - @globalfishingwatch/dataviews-client@11.8.0
+  - @globalfishingwatch/layer-composer@8.10.0
+
+## 0.0.71
+
+### Patch Changes
+
+- Updated dependencies [0ea624c9]
+- Updated dependencies [08b8209f]
+  - @globalfishingwatch/layer-composer@8.9.1
+  - @globalfishingwatch/api-client@2.0.1
+
 ## 0.0.70
 
 ### Patch Changes