+++ packages/ocean-areas/CHANGELOG.md
@@ -1,5 +1,23 @@
 ## 0.0.1
 
+## 0.2.5
+
+### Patch Changes
+
+- fix chinese taipei label
+
+## 0.2.4
+
+### Patch Changes
+
+- cc73f7d3: fix Chinese Taipei label
+
+## 0.2.3
+
+### Patch Changes
+
+- 52296801: fix Falkland label
+
 ## 0.2.2
 
 ### Patch Changes